// types/capabilities.ts
export enum Capability {
    onoff = "onoff",
    dim = "dim",
    light_hue = "light_hue",
    light_saturation = "light_saturation",
    light_temperature = "light_temperature",
    light_mode = "light_mode",
    vacuumcleaner_state = "vacuumcleaner_state",
    thermostat_mode = "thermostat_mode",
    target_temperature = "target_temperature",
    measure_temperature = "measure_temperature",
    measure_co = "measure_co",
    measure_co2 = "measure_co2",
    measure_pm25 = "measure_pm25",
    measure_humidity = "measure_humidity",
    measure_pressure = "measure_pressure",
    measure_noise = "measure_noise",
    measure_rain = "measure_rain",
    measure_wind_strength = "measure_wind_strength",
    measure_wind_angle = "measure_wind_angle",
    measure_gust_strength = "measure_gust_strength",
    measure_gust_angle = "measure_gust_angle",
    measure_battery = "measure_battery",
    measure_power = "measure_power",
    measure_voltage = "measure_voltage",
    measure_current = "measure_current",
    measure_luminance = "measure_luminance",
    measure_ultraviolet = "measure_ultraviolet",
    measure_water = "measure_water",
    alarm_generic = "alarm_generic",
    alarm_motion = "alarm_motion",
    alarm_contact = "alarm_contact",
    alarm_co = "alarm_co",
    alarm_co2 = "alarm_co2",
    alarm_pm25 = "alarm_pm25",
    alarm_tamper = "alarm_tamper",
    alarm_smoke = "alarm_smoke",
    alarm_fire = "alarm_fire",
    alarm_heat = "alarm_heat",
    alarm_water = "alarm_water",
    alarm_battery = "alarm_battery",
    alarm_night = "alarm_night",
    meter_power = "meter_power",
    meter_water = "meter_water",
    meter_gas = "meter_gas",
    meter_rain = "meter_rain",
    homealarm_state = "homealarm_state",
    volume_set = "volume_set",
    volume_up = "volume_up",
    volume_down = "volume_down",
    volume_mute = "volume_mute",
    channel_up = "channel_up",
    channel_down = "channel_down",
    locked = "locked",
    lock_mode = "lock_mode",
    garagedoor_closed = "garagedoor_closed",
    windowcoverings_state = "windowcoverings_state",
    windowcoverings_tilt_up = "windowcoverings_tilt_up",
    windowcoverings_tilt_down = "windowcoverings_tilt_down",
    windowcoverings_tilt_set = "windowcoverings_tilt_set",
    windowcoverings_closed = "windowcoverings_closed",
    windowcoverings_set = "windowcoverings_set",
    button = "button",
    speaker_playing = "speaker_playing",
    speaker_next = "speaker_next",
    speaker_prev = "speaker_prev",
    speaker_shuffle = "speaker_shuffle",
    speaker_repeat = "speaker_repeat",
    speaker_artist = "speaker_artist",
    speaker_album = "speaker_album",
    speaker_track = "speaker_track",
    speaker_duration = "speaker_duration",
    speaker_position = "speaker_position"
}