// types/classes.ts
export enum Class {
    amplifier = "amplifier",
    blinds = "blinds",
    button = "button",
    camera = "camera",
    coffeemachine = "coffeemachine",
    curtain = "curtain",
    doorbell = "doorbell",
    fan = "fan",
    garagedoor = "garagedoor",
    heater = "heater",
    homealarm = "homealarm",
    kettle = "kettle",
    light = "light",
    lock = "lock",
    other = "other",
    remote = "remote",
    sensor = "sensor",
    socket = "socket",
    speaker = "speaker",
    solarpanel = "solarpanel",
    sunshade = "sunshade",
    thermostat = "thermostat",
    tv = "tv",
    vacuumcleaner = "vacuumcleaner",
    windowcoverings = "windowcoverings"
}